Improve usage of SelectStatementVisitor
It now has a new constructor that takes a set as an argument. The constructor
checks whether the set can be used (i.e. is not null and is empty). This way,
anyone using the visitor can no longer forget to pass along a set to use for
the rules generated by the visitor.